feature: relation manager de producto-articulo
