+++ app/Filament/Resources/ArticuloResource.php
@@ -3,6 +3,8 @@
 namespace App\Filament\Resources;
 
 use App\Filament\Resources\ArticuloResource\Pages;
+use App\Filament\Resources\ArticuloResource\RelationManagers\ArticulosRelationManager;
+use App\Filament\Resources\ArticuloResource\Traits\ArticuloForms;
 use App\Helpers\ClaveGenerator;
 use App\Helpers\Helpers;
 use App\Models\Articulo;
@@ -33,6 +35,8 @@ use Livewire\Features\SupportFileUploads\TemporaryUploadedFile;
 
 class ArticuloResource extends Resource
 {
+    use ArticuloForms;
+
     protected static ?string $model = Articulo::class;
 
     protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
@@ -45,142 +49,11 @@ class ArticuloResource extends Resource
                     ->tabs([
                         Tabs\Tab::make('Datos')
                             ->schema([
-                                Grid::make()
-                                    ->columns(12)
-                                    ->schema([
-
-                                        TextInput::make('nombre')
-                                            ->columnSpan([
-                                                'md' => 8,
-                                            ])
-                                            ->required(),
-                                        Toggle::make('insumo')
-                                            ->columnSpan([
-                                                'md' => 4,
-                                            ])
-                                            ->hint('¿El artículo es un insumo?')
-                                            ->columnSpanFull(),
-                                        TextInput::make('clave')
-                                            ->columnSpan([
-                                                'md' => 8,
-                                            ])
-                                            ->required()
-                                            ->hintAction(
-                                                Action::make('generarClaves')
-                                                    ->icon('heroicon-m-cog-6-tooth')
-                                                    ->action(function (string|float|null $state, Set $set, Get $get, Livewire $livewire) {
-                                                        $claves = ClaveGenerator::generarClaves($get('nombre'));
-                                                        $set('clave', $claves['clave']);
-                                                        $set('clave_alterna', $claves['clave_alterna']);
-                                                    })
-                                            ),
-                                        TextInput::make('clave_alterna')
-                                            ->columnSpan([
-                                                'md' => 4,
-                                            ])
-                                            ->required(),
-                                        Fieldset::make('Producto')
-                                            ->schema([
-                                                Select::make('producto_id')
-                                                    ->relationship(name: 'producto', titleAttribute: 'nombre')
-                                                    ->columnSpan([
-                                                        'md' => 8,
-                                                    ])
-                                                    ->required()
-                                                    ->searchable(['nombre'])
-                                                    ->preload()
-                                                    ->afterStateUpdated(function (string|float|null $state, Set $set, Get $get) {
-                                                        $record = Producto::find($state);
-                                                        $set('costo_unitario', $record->costo_unitario ?? 0);
-                                                        $set('precio', Helpers::makePrecio(
-                                                            $record->costo_unitario ?? 0,
-                                                            $get('valor_equivalente'),
-                                                            $get('porcentaje')
-                                                        ));
-                                                    })
-                                                    ->live(onBlur: true),
-                                                TextInput::make('valor_equivalente')
-                                                    ->columnSpan([
-                                                        'md' => 4,
-                                                    ])
-                                                    ->required()
-                                                    ->numeric()
-                                                    ->minValue(0)
-                                                    ->default(0)
-                                                    ->prefixIcon('heroicon-o-inbox-stack')
-                                                    ->hintIcon('heroicon-m-question-mark-circle', tooltip: 'Ejemplo: Equivale a 4 litros de producto X.'),
-                                            ])
-                                            ->columns(12),
-                                        Fieldset::make('Costos')
-                                            ->schema([
-                                                TextInput::make('costo_unitario')
-                                                    ->numeric()
-                                                    ->minValue(0)
-                                                    ->prefix('$')
-                                                    ->disabled()
-                                                    ->readOnly(),
-                                                TextInput::make('porcentaje')
-                                                    ->numeric()
-                                                    ->prefix('%')
-                                                    ->minValue(0)
-                                                    ->default(1)
-                                                    ->step(0.1),
-                                                TextInput::make('precio')
-                                                    ->numeric()
-                                                    ->required()
-                                                    ->prefix('$')
-                                                    ->minValue(0)
-                                                    ->default(0)
-                                                    ->step(0.01)
-                                                    ->hintIcon('heroicon-m-question-mark-circle', tooltip: 'Precio del artículo.')
-                                                    ->suffixAction(
-                                                        Action::make('generarPrecio')
-                                                            ->icon('heroicon-m-sparkles')
-                                                            ->action(function (string|float|null $state, Set $set, Get $get) {
-                                                                $set('precio',  Helpers::makePrecio(
-                                                                    $get('costo_unitario'),
-                                                                    $get('valor_equivalente'),
-                                                                    $get('porcentaje')
-                                                                ));
-                                                            })
-                                                    ),
-                                                Toggle::make('usa_embace')
-                                                    ->columnSpanFull()
-                                                    ->helperText('¿El artículo usa embace?')
-                                                    ->live(),
-                                                TextInput::make('precio_embase')
-                                                    ->numeric()
-                                                    ->prefix('$')
-                                                    ->minValue(0)
-                                                    ->step(0.1)
-                                                    ->requiredIf('usa_embace', true)
-                                                    ->columnSpanFull()
-                                                    ->hidden(fn (Get $get): bool => !$get('usa_embace')),
-                                            ])
-                                            ->columns(3),
-                                        Textarea::make('descripcion')
-                                            ->translateLabel()
-                                            ->columnSpanFull(),
-
-                                    ])
+                                self::mainform()
                             ]),
                         Tabs\Tab::make('Imagenes')
                             ->schema([
-                                FileUpload::make('imagenes')
-                                    ->acceptedFileTypes(['image/jpeg', 'image/png', 'image/webp'])
-                                    ->multiple()
-                                    ->imageEditor()
-                                    ->directory(
-                                        fn (?Articulo $record): string => Articulo::IMAGE_DIRECTORY . $record->id
-                                    )
-                                    ->visibility('public')
-                                    ->reorderable()
-                                    ->getUploadedFileNameForStorageUsing(
-                                        fn (TemporaryUploadedFile $file): string => (string) str($file->getClientOriginalName())
-                                            ->prepend('articulo-image' . str()->uuid()),
-                                    )
-                                    ->maxFiles(4)
-                                    ->columnSpanFull(),
+                                self::imageForm()
                             ])
                             ->visible(fn (?Articulo $record) => $record !== null)
 
@@ -212,32 +85,11 @@ class ArticuloResource extends Resource
                         TextColumn::make('nombre')
                             ->searchable()
                             ->sortable()
-                            ->description(function (Articulo $articulo) {
-                                return $articulo?->clave;
-                            })
                             ->columnSpanFull(),
-
+                        TextColumn::make('clave')
+                            ->searchable()
+                            ->columnSpanFull()
                     ]),
-
-
-                /* TextColumn::make('clave')
-                    ->searchable()
-                    ->sortable()
-                    ->description(fn (Articulo $articulo) => $articulo?->clave_alterna),
-                TextColumn::make('valor_equivalente')
-                    ->searchable()
-                    ->sortable(),
-                TextColumn::make('precio')
-                    ->searchable()
-                    ->sortable()
-                    ->formatStateUsing(fn (Articulo $articulo) => $articulo->precioFormat),
-                IconColumn::make('usa_embace')
-                    ->boolean()
-                    ->sortable(),
-                TextColumn::make('insumo')
-                    ->badge()
-                    ->formatStateUsing(fn (Articulo $articulo) => $articulo->esInsumo)
-                    ->color(fn ($state): string => $state ? 'success' : 'primary') */
             ])
             ->filters([
                 //
@@ -256,9 +108,7 @@ class ArticuloResource extends Resource
 
     public static function getRelations(): array
     {
-        return [
-            //
-        ];
+        return [];
     }
 
     public static function getPages(): array
